+++ packages/visual-editor/src/wdyr.ts
@@ -1,0 +1,5 @@
+import React from 'react';
+import whyDidYouRender from '@welldone-software/why-did-you-render';
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+
+whyDidYouRender(React, {});