+++ packages/experience-builder-sdk/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0-alpha.10](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder@4.0.0-alpha.9...@contentful/experience-builder@4.0.0-alpha.10) (2024-01-11)
+
+**Note:** Version bump only for package @contentful/experience-builder
+
+# [4.0.0-alpha.9](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder@4.0.0-alpha.8...@contentful/experience-builder@4.0.0-alpha.9) (2024-01-11)
+
+### Bug Fixes
+
+- **experience-builder-sdk:** export ComponentDefintion ([#243](https://github.com/contentful/experience-builder/issues/243)) ([b74d292](https://github.com/contentful/experience-builder/commit/b74d2927575cc10dcd4bb7e4800c47b979d7e15b))
+
+# [4.0.0-alpha.8](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder@4.0.0-alpha.7...@contentful/experience-builder@4.0.0-alpha.8) (2024-01-11)
+
+### Bug Fixes
+
+- invalid calc value for width CSS property ([65dc14c](https://github.com/contentful/experience-builder/commit/65dc14c1aa68038fc9f5152c1c6aa270474b1392))
+- **visual-editor:** render assembly nodes by passing the required props [ALT-255] ([#245](https://github.com/contentful/experience-builder/issues/245)) ([6d2dd5a](https://github.com/contentful/experience-builder/commit/6d2dd5a64337e13ac84c0014da2c03e519f7f78b))
+
 # [4.0.0-alpha.7](https://github.com/contentful/experience-builder/compare/@contentful/experience-builder@4.0.0-alpha.6...@contentful/experience-builder@4.0.0-alpha.7) (2024-01-10)
 
 **Note:** Version bump only for package @contentful/experience-builder